requirements.txt
# Core dependencies for Chronos Engine

# YAML parsing across modules and commands
PyYAML>=6.0

# Audio playback for alarms/reminders via pygame.mixer
pygame>=2.5.0

# Optional console color support (some commands may leverage it)
colorama>=0.4.6

# Optional simple audio utility; retained for compatibility
playsound>=1.2.2

# CLI UX (autocomplete/suggestions)
prompt_toolkit>=3.0.43

# Desktop tray app + icon rendering
pystray>=0.19.5
Pillow>=10.3.0

# Embedded webview windows for Topos
pywebview

